renamed from startdialog to startpreservationdialog
